MAINT: Fixed performance issue with rfft2 and irfft2
When fftn or ifftn are called on 1D arrays we should go though fft/ifft, instead
of using _direct_fftnd to take advantage of caching.

Unlike in rfftn_numpy, where out-of-place rfft_numpy is performed first, and
the subsequent fftn calls can be in-place, for irfftn_numpy the ifftn need to be
performance first, and one must check whether the first iteration can in fact
be performed in place.