BUGFIX: Issue #21
When computing parameters for multiple equidistant datasets call to MKL's FFT functions,
during computation of distances between datasets there was an implicit assumption that
strides are all positive.

The code was caught off-guards for arrays with a zero stride, associated with a
unit shape, like one formed by a[np.newaxis]

Added a test.