BUG: Fixed issue/37
See https://github.com/IntelPython/mkl_fft/issues/37

As noted by the reported, the logic of _remove_axes was not right.
We always have to remove the last axis (since _cook_nd_args rearranges
it that way), but upon remove of the axis, we have to decrease all
the axis higher than the pivot by one, since we take part at the pivot.